dm-verity: Revert
Enable dm-verity on flounder

Enable dm-verity for system and vendor partitions. Store the state in
the crypto metadata partition, and display a warning if dm-verity is
started in logging mode.

Adjusting read-ahead and prefetch_cluster settings for the dm-verity
devices reduces overhead to ~1% for sequential reads with block sizes
from 1kB to 32kB, and to ~9% for 512 byte blocks, when compared to the
best configuration for the raw block device.

Change-Id: Ia01ebab787ed65f52c7fb28abd4e0e2814bce97a

DO NOT MERGE: disable image signing

Do not sign boot or recovery images with boot_signer during the build
process, and do not append signed dm-verity metadata to system or vendor
images.

Bug: 19985143
Change-Id: If2cc4ecdfd9ed9984e94769d9501dafbae71ba21
Signed-off-by: Jean-Pierre Rasquin <yank555.lu@gmail.com>